pyproject.toml
[project]
name = "fraud-detection-ml-system"
version = "0.1.0"
description = "Production-grade fraud detection ML service"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "mlflow>=2.11",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "omegaconf>=2.3",
    "joblib>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.9",
    "pre-commit>=3.7",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
fraud-serve = "src.api.main:main"

[tool.setuptools]
packages = ["src", "src.utils", "src.ingestion", "src.validation", "src.features", "src.training", "src.evaluation", "src.registry", "src.api"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "N"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true